move: GoTopBtn 컴포넌트 위치 수정
